+++ src/heuristics/mod.rs
@@ -3,6 +3,9 @@
 //! The templates in this module allow customization of certain operators
 //! similar to other frameworks (e.g. a custom mutation operator) and are meant to
 //! provide a starting point for more specialized configurations.
+//!
+//! A common workflow is to copy the definition of the heuristic closest to the desired
+//! structure and adapt it to one's needs.
 
 pub mod aco;
 pub mod cro;
